requirements.txt
gradio>=3.50.0
langchain>=0.0.300
langchain-openai>=0.0.2
langchain-community>=0.0.20
openai>=1.0.0
groq>=0.9.0
huggingface_hub>=0.19.0
PyMuPDF>=1.23.0
chromadb>=0.4.0
loguru>=0.7.0
cryptography>=41.0.0
tiktoken>=0.5.0
scikit-learn>=1.3.0
python-Levenshtein>=0.20.0
numpy>=1.26.0
python-dotenv>=1.0.0
moviepy>=1.0.3
markdown-pdf>=1.1.2
Pillow>=10.0.0
speechrecognition>=3.10.0